project/requirements.txt
google-generativeai>=0.5.0
numpy>=1.24.0
pathlib2>=2.3.7
